Procedures commands refactored to keep them isolated from payments commands.
