--- app/commands/register_membership_level_change.rb
@@ -1,29 +1,0 @@
-# frozen_string_literal: true
-
-# A command to register a change of membership for a person.
-class RegisterMembershipLevelChange < Rectify::Command
-  # Public: Initializes the command.
-  #
-  # person - A registered person that want to change its level of membership.
-  # to_level - The desired level of membership for the person
-  def initialize(person, to_level)
-    @person = person
-    @to_level = to_level
-  end
-
-  # Executes the command. Broadcasts these events:
-  #
-  # - :ok when everything is valid.
-  # - :invalid if the procedure wasn't valid and we couldn't proceed.
-  #
-  # Returns nothing.
-  def call
-    result = :invalid
-    if @person.level != @to_level
-      membership_level_change = Procedures::MembershipLevelChange.new(person: @person, from_level: @person.level, to_level: @to_level)
-      result = :ok if membership_level_change.save
-    end
-
-    broadcast result
-  end
-end
