+++ app/controllers/api/v1/people_controller.rb
@@ -16,7 +16,7 @@ module Api
 
     def change_membership_level
       @person = Person.find_by("extra ->> 'participa_id' = ?", params[:id])
-      RegisterMembershipLevelChange.call(@person, params[:level]) do
+      Procedures::RegisterMembershipLevelChange.call(@person, params[:level]) do
         on(:invalid) do
           render json: @person.errors, status: :unprocessable_entity
         end
